Seperated input and list items into seperate components
